refactor(rye-app): extract WorkerUi (worker-side egui integration) to wasm/worker_ui.rs (was inlined in worker.rs)
